make URI integers 16 bits, fixes #5443

+++ pdns/dnsrecords.hh
@@ -675,7 +675,7 @@ class URIRecordContent : public DNSRecordContent {
   public:
     includeboilerplate(URI)
   private:
-    uint8_t d_priority, d_weight;
+    uint16_t d_priority, d_weight;
     string d_target;
 };
